extra/requirements.txt
dotenv
mysql-connector-python 
Fakers